+++ src/engine/engine_io.c
@@ -438,6 +438,9 @@ mjModel* mj_makeModel(int nq, int nv, int nu, int na, int nbody, int njnt,
   m->nuser_actuator = nuser_actuator;
   m->nuser_sensor = nuser_sensor;
   m->nnames = nnames;
+  m->nnames_map = nbody + njnt + ngeom + nsite + ncam + nlight + nmesh + nskin
+                  + nhfield + ntex + nmat + npair + nexclude + neq + ntendon
+                  + nu + nsensor + nnumeric + ntext + ntuple + nkey + nplugin;
 
 #define X(name)                                    \
   if ((m->name) < 0) {                             \