+++ src/engine/engine_support.h
@@ -78,6 +78,9 @@ MJAPI int mj_jacDifPair(const mjModel* m, const mjData* d, int* chain,
 
 //-------------------------- name functions --------------------------------------------------------
 
+// get string hash, see http://www.cse.yorku.ca/~oz/hash.html
+uint64_t mj_hashdjb2(const char* s, uint64_t n);
+
 // get id of object with specified name; -1: not found; type is mjtObj
 MJAPI int mj_name2id(const mjModel* m, int type, const char* name);